fix: split scheduled edition cron routes
